+++ lib/action_reporter/error.rb
@@ -1,0 +1,3 @@
+module ActionReporter
+  class Error < StandardError; end
+end