+++ lib/action_reporter/sentry_reporter.rb
@@ -3,23 +3,27 @@ module ActionReporter
     class_accessor "Sentry"
 
     def notify(error, context: {})
-      Sentry.with_scope do |temp_scope|
+      sentry_class.with_scope do |temp_scope|
         temp_scope.set_context("context", transform_context(context))
 
         if error.is_a?(StandardError)
-          Sentry.capture_exception(error)
+          sentry_class.capture_exception(error)
         else
-          Sentry.capture_message(error)
+          sentry_class.capture_message(error)
         end
       end
     end
 
     def context(args)
-      Sentry.get_current_scope.set_context("context", transform_context(args))
+      sentry_class.get_current_scope.set_context("context", transform_context(args))
+    end
+
+    def reset_context
+      sentry_class.get_current_scope.set_context("context", {})
     end
 
     def audited_user=(user)
-      Sentry.set_user(global_id: user.to_global_id.to_s) if user
+      sentry_class.set_user(user_global_id: user&.to_global_id&.to_s)
     end
   end
 end