+++ spec/action_reporter/audited_reporter_spec.rb
@@ -3,10 +3,23 @@ require 'rails'
 require 'audited'
 
 RSpec.describe ActionReporter::AuditedReporter do
-  subject { described_class.new }
+  subject(:instance) { described_class.new }
 
-  let(:audited_stub) { double('Audited') }
-  before do
-    stub_const('Audited', audited_stub)
+  describe '#notify' do
+
+  end
+
+  describe '#context' do
+
+  end
+
+  describe '#reset_context' do
+    subject(:reset_context) { instance.reset_context }
+
+    it 'resets context' do
+      expect(Audited.store).to receive(:delete).with(:current_remote_address)
+      expect(Audited.store).to receive(:delete).with(:audited_user)
+      subject
+    end
   end
 end