+++ spec/action_reporter/honeybadger_reporter_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'honeybadger'
 
 RSpec.describe ActionReporter::HoneybadgerReporter do
-  subject { described_class.new }
+  subject(:instance) { described_class.new }
 
   before do
     Honeybadger.configure do |config|
@@ -11,15 +11,56 @@ RSpec.describe ActionReporter::HoneybadgerReporter do
   end
 
   describe '#notify' do
+    subject(:notify) { instance.notify(error, context: context_data) }
 
+    let(:error) { StandardError.new('error') }
+    let(:context_data) { { foo: 'bar' } }
+
+    it 'captures error' do
+      expect(Honeybadger).to receive(:notify).with(error, { context: context_data }).and_call_original
+      subject
+    end
   end
 
   describe '#context' do
+    subject(:context) { instance.context(context_data) }
 
+    let(:context_data) { { foo: 'bar' } }
+
+    it 'sets context' do
+      expect(Honeybadger).to receive(:context).with(context_data).and_call_original
+      subject
+    end
+
+    it 'transforms context' do
+      expect(instance).to receive(:transform_context).with(context_data).and_call_original
+      subject
+    end
+  end
+
+  describe '#reset_context' do
+    subject(:reset_context) { instance.reset_context }
+
+    it 'resets context' do
+      expect(Honeybadger.context).to receive(:clear!).and_call_original
+      subject
+    end
+  end
+
+  describe '#audited_user=' do
+    subject(:audited_user=) { instance.audited_user = user }
+
+    let(:sample_id) { double("GlobalId", to_s: "user-global-id") }
+    let(:user) { double("User", to_global_id: sample_id) }
+
+    it 'sets user_global_id' do
+      expect(Honeybadger).to receive(:context).with(user_global_id: sample_id.to_s).and_call_original
+      subject
+    end
   end
 
   describe '#check_in' do
-    subject(:check_in) { described_class.new.check_in(identifier) }
+    subject(:check_in) { instance.check_in(identifier) }
 
     context "when identifier is a string" do