+++ spec/action_reporter/sentry_reporter_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'sentry-ruby'
 
 RSpec.describe ActionReporter::SentryReporter do
-  subject { described_class.new }
+  subject(:instance) { described_class.new }
 
   before do
     Sentry.init do |config|
@@ -11,31 +11,62 @@ RSpec.describe ActionReporter::SentryReporter do
   end
 
   describe '#notify' do
-    context "when error is an exception" do
-      let(:error) { StandardError.new('Error') }
+    subject(:notify) { instance.notify(error, context: context) }
 
-      it 'sends notification' do
-        # expect(Sentry).to receive(:get_current_scope).and_call_original
-        expect(Sentry).to receive(:capture_exception).with(error).and_call_original
-        subject.notify(error, context: { foo: 'bar' })
-      end
+    let(:error) { StandardError.new('Error') }
+    let(:context) { { foo: 'bar' } }
+
+    it 'sends notification' do
+      expect(Sentry).to receive(:with_scope).and_call_original
+      expect(Sentry).to receive(:capture_exception).with(error).and_call_original
+      subject
     end
 
     context "when error is a string" do
       let(:error) { 'Error' }
 
       it 'sends notification' do
-        # expect(Sentry).to receive(:get_current_scope).and_call_original
+        expect(Sentry).to receive(:with_scope).and_call_original
         expect(Sentry).to receive(:capture_message).with(error).and_call_original
-        subject.notify(error, context: { foo: 'bar' })
+        subject
       end
     end
   end
 
   describe '#context' do
+    subject(:context) { instance.context(context_data) }
+
+    let(:context_data) { { foo: 'bar' } }
+
     it 'sets context' do
-      expect(Sentry).to receive(:get_current_scope).and_call_original
-      subject.context({ foo: 'bar' })
+      expect(Sentry.get_current_scope).to receive(:set_context).with("context", context_data).and_call_original
+      subject
+    end
+
+    it 'transforms context' do
+      expect(instance).to receive(:transform_context).with(context_data).and_call_original
+      subject
+    end
+  end
+
+  describe '#reset_context' do
+    subject(:reset_context) { instance.reset_context }
+
+    it 'resets context' do
+      expect(Sentry.get_current_scope).to receive(:set_context).with("context", {}).and_call_original
+      subject
+    end
+  end
+
+  describe '#audited_user=' do
+    subject(:audited_user=) { instance.audited_user = user }
+
+    let(:sample_id) { double("GlobalId", to_s: "user-global-id") }
+    let(:user) { double("User", to_global_id: sample_id) }
+
+    it 'sets user_global_id' do
+      expect(Sentry).to receive(:set_user).with(user_global_id: sample_id.to_s).and_call_original
+      subject
     end
   end
 end